backend/requirements.txt
python-telegram-bot[job-queue,rate-limiter]>=21.0
supabase>=2.0.0
python-dotenv>=1.0.0
